+++ e2e/wdio/platform/pages/time-picker.po.ts
@@ -16,8 +16,11 @@ export class TimePickerPO extends BaseComponentPo {
     disabledInput = this.disabledTimePicker + ' input';
     disabledButton = this.disabledTimePicker + ' button';
     navigationDownArrowButton = 'button[glyph="navigation-down-arrow"]';
-    timeItem = 'span.fd-time__item';
+    timeColumn = 'fd-time-column';
     selectedValue = 'div.fd-time__current-indicator ~ span';
+    hoursColumn = '.fd-time .fd-time__col:nth-child(1) ';
+    thirdColumn = '.fd-time .fd-time__col:nth-child(3) ';
+    columnItem = '.fd-time__item:not(.fd-time__item--collapsed)';
 
     getScreenshotFolder(): Record<string, any> {
         return super.getScreenshotFolder(this.url);