+++ libs/platform/src/lib/form/form-generator/interfaces/dynamic-form-group.ts
@@ -1,9 +1,9 @@
 import { FormGroup } from '@angular/forms';
 
-import { DynamicFormControl } from '../dynamic-form-control';
+import { DynamicFormControl, DynamicFormControlGroup } from '../dynamic-form-control';
 
 export interface DynamicFormGroup extends FormGroup {
     controls: {
-        [key: string]: DynamicFormControl;
+        [key: string]: DynamicFormControl | DynamicFormControlGroup;
     };
 }